+++ src/main/java/com/climbing/auth/login/GetLoginMember.java
@@ -1,4 +1,4 @@
-package com.climbing.domain.member;
+package com.climbing.auth.login;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
