+++ src/main/java/com/climbing/auth/oauth2/SocialType.java
@@ -1,4 +1,4 @@
-package com.climbing.domain.member;
+package com.climbing.auth.oauth2;
 
 public enum SocialType {
     GOOGLE, NAVER, KAKAO
